requirements.txt
anyio==3.6.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.85.2
fastapi-utils==0.2.1
Flask==2.2.2
Flask-Cors==3.0.10
greenlet==1.1.3.post0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
Pillow==9.3.0
psycopg2-binary==2.9.3
pydantic==1.10.2
PyJWT==2.6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
torch==1.13.0
torchvision==0.14.0
typing-extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.9.0
